add partial context and modal context
